requirements.txt
streamlit==1.32.0
supabase==2.3.5
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
pydantic>=2.4.0
uuid>=1.30
pillow>=10.0.0 